add a anime

+++ routes/web.php
@@ -40,10 +40,10 @@ Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name
 
 // Admin only
     Route::get('/admin-pannel', [DashboardController::class, 'index'])->middleware(['auth', 'isAdmin']);
+
     Route::get('/users', [DashboardController::class, 'userIndex'])->name('user.index')->middleware(['auth', 'isAdmin']);
-    Route::delete('/users/{user}', [DashboardController::class, 'userDelete'])->name('userDelete');
+    Route::delete('/users/{user}', [DashboardController::class, 'userDelete'])->name('userDelete')->middleware(['auth', 'isAdmin']);
+    Route::put('/users/{user}', [DashboardController::class, 'userUpdate'])->name('userUpdate')->middleware(['auth', 'isAdmin']);
 
-    
+    Route::post('/admin-pannel', [DashboardController::class, 'animeCreate'])->name('animeCreate')->middleware(['auth', 'isAdmin']);
 
-
-// });
